Add RequestProcessor CRD
The RequestProcessor is able to expose an application or function build
via a Knative Serving. Each item create a Knative Configuration
including either the specified image or the image from either a
referenced ApplicationBuild or FunctionBuild or created from an inlined
ApplicationSpec or BuildSpec.

A single Knative Route is created to route traffic to each item. Unlike
Knative Routes, each item will receive an equal amount of traffic that
is not otherwise specified. Any configuration that should not receive
traffic should explicitly set the percentage to zero. Tagged items are
also exposed as named entries in the route enabling direct access
independent of the routes percentage.

Note: because traffic is defined as an integer that must sum exactly to
100%, items that do not explicitly specify a percentage may receive 1%
less traffic than another. An item may also receive no traffic. Items
defined earlier in the list will receive the remainder of traffic when
the split is not evenly divisible.

- * => 100
- *,* => 50,50
- *,*,* => 34,33,33
- *,1 => 99,1
- 99,*,* => 99,1,0