Require RequestProcessor items be named (#12)
The RequestProcessor spec is an array, to faciliate consistent
reconciliation each item needs a stable key. Otherwise reordering or
inserting a new item could cause a running configuration to be replaced
when it isn't nessesary to do so. Moreover, without a consistent key,
the Revision history for each Configuration is meaningless.

Each item now must be uniquely named. Configurations with the same name
will always be updated temporally in an unambiguous way. Renaming a
configuration is equivlent to a drop and add rather than an update.
Reordering configurations is a no-op.

The name also gives tooling something meaningful to anchor changes to
rather than using an index.

Also in this change is no longer inlining the PodSpec, it is now called
Template. This will eliminate field collisions between the item and pod.