Reconcile credentials to riff-build ServiceAccount (#15)
Credentials are Secrets with a label marking them as a riff Credential.
The credentails should be bound to the riff-build service account in the
namespace the secrets exist. When a credential is deleted, the service
account should no longer contain a reference to that secret.

Other non-credential secrets are not impacting durring these operations.
An annotation is added to the service account to indicate which secrets
are credentials.