Release script (#18)
For successful builds on master, build and publish image and yaml to GCR
and GCS.

Images are published to `gcr.io/cf-spring-pfs-eng/riff-system/...` and
tags with the project version and the version with the curent git commit
sha. Config files are published to `gs://projectriff/riff-system/` with
the current version and to `gs://projectriff/riff-system/snapshots` with
the version and gitsha.