Convert request handler to use k8s types (#41)
Introducing the core runtime which resolves a Handler to basic Kubernetes Deployments and Services. This runtime does not offer ingress or autoscaler, however, there are other projects and components that can manage that role, like Ingress and the HorizontalPodAutoscaler.

Users have freedom to define how these resources will be exposed and managed.