Run FATS tests between stage and publish CI jobs (#97)
Tests the core and knative runtimes in isolation by building and
invoking both a function and application.

Due to the build vm being limited to 2 CPUs we use ytt to remove Pod
resource requests and limits. We could skip this step if running on a
bigger cluster, but that would required credentials that are not
available for PRs.