Remove namespace definition in runtimes (#175)
The namespace is defined in the build system. It does not need to be
redefined in each runtime. For example, the helm charts jump through
hoops to remove the duplicate namespaces. It also make `kubectl delete`
dangerous to delete a runtime, as it will remove all controllers when
the namespace is removed.