Move health checks to separate port (#185)
controller-runtime 0.4 now has support for hosting health checks. We no
longer need to attempt to connect to the webhook port. This eliminates
TLS handshake errors in the logs which were triggered by the probe.

"Default" checks are registered, which will always report back as
healthy. We can create more nuanced checks as needed.