Allow custom container port (#187)
Previously the runtime assumed port 8080. Now the port may be defined in
the container's ContainerPorts list. The containers first port is used
when defaulting probes and creating the fronting service mapping.

The PORT envvar is now set on the first container of the Deployment.
This envvar is prohibited from being set directly on the Deployer as it
will be injected.

8080 is still used by default.

Resolves #48