Introduce IngressPolicy on Deployers (#191)
The IngressPolicy defines whether the workload is exposed outside the
cluster via ingress. There are two policies:

- `ClusterLocal`: no ingress
- `External`: creates ingress

`External ` is the default policy, which preserves the current behavior.

Note: the cluster local policy is not sufficient to protect workload
from external network access if the cluster is also not adequately
protected. It merely prevents the workload from being registered via
kubernetes ingress.

Resolves #179