Update k8s to 1.16.4
Starting with 1.16.4, releases use semantic tags that are compatible
with go modules. One quick is that the major version is listed as `v0`,
however, the v0.16.4 and kubernetes-1.16.4 tags match.

This change means that Dependabot will start tracking these dependencies
for us, although they should all be upgraded in a group.