Drop liveness probe (#276)
Generic liveness probes can be problematic when the performance
characteristics of a container don't align with the default values. For
example, a slow to start container can be marked as down and terminated
before it can finish starting. This will place the container into a
permanent crash loop.

Users may still define a custom liveness probe, and a default readiness
probe is still provided.